tckglobal: Fix linux build warnings.

+++ modules/gt/src/particlegrid.cpp
@@ -78,9 +78,9 @@ namespace MR {
           list.clear();
         }
         
-        const ParticleGrid::ParticleVectorType* ParticleGrid::at(const int x, const int y, const int z) const
+        const ParticleGrid::ParticleVectorType* ParticleGrid::at(const ssize_t x, const ssize_t y, const ssize_t z) const
         {
-          if ((x < 0) || (x >= dims[0]) || (y < 0) || (y >= dims[1]) || (z < 0) || (z >= dims[2]))  // out of bounds
+          if ((x < 0) || (size_t(x) >= dims[0]) || (y < 0) || (size_t(y) >= dims[1]) || (z < 0) || (size_t(z) >= dims[2]))  // out of bounds
             return nullptr;
           return &grid[xyz2idx(x, y, z)];
         }
